requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0

# ── Scraper pipeline ────────────────────────────────────────────────
instaloader>=4.10
opencv-python>=4.8.0
yt-dlp>=2024.0
playwright>=1.40